third_party/blink/web_tests/http/tests/history/post-replace-state-reload.html

<body>
<form method="post"><input type="submit" value="submit"></input></form>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpFrameLoadCallbacks();
    testRunner.waitUntilDone();
}

history.replaceState("", "");

window.onload = function() {
    if (sessionStorage.getItem("post-replace-reload") == null) {
        sessionStorage.setItem("post-replace-reload", "submit");
        document.forms[0].submit();
    } else if (sessionStorage.getItem("post-replace-reload") == "submit") {
        sessionStorage.setItem("post-replace-reload", "reload");
        if (window.internals)
            internals.forceReload(false);
    } else {
        sessionStorage.removeItem("post-replace-reload");
        if (window.testRunner)
            testRunner.notifyDone();
    }
}
</script>
</body>
